Move Username/Password login form to a new component and prepare for adding 2FA
