+++ client/blocks/login/test/login-form.jsx
@@ -14,19 +14,19 @@ import FormTextInput from 'components/forms/form-text-input';
 import FormPasswordInput from 'components/forms/form-password-input';
 import FormsButton from 'components/forms/form-button';
 
-describe( 'Login', function() {
-	let Login;
+describe( 'LoginForm', function() {
+	let LoginForm;
 
 	useFakeDom();
 
 	before( () => {
-		Login = require( 'blocks/login' ).Login;
+		LoginForm = require( 'blocks/login/login-form' ).LoginForm;
 	} );
 
 	context( 'component rendering', () => {
 		it( 'displays a login form', () => {
 			const wrapper = shallow(
-				<Login translate={ noop } />
+				<LoginForm translate={ noop } />
 			);
 			expect( wrapper.find( FormTextInput ).length ).to.equal( 1 );
 			expect( wrapper.find( FormPasswordInput ).length ).to.equal( 1 );
@@ -35,7 +35,7 @@ describe( 'Login', function() {
 
 		it( 'shows the header text', () => {
 			const wrapper = shallow(
-				<Login
+				<LoginForm
 					translate={ noop }
 					title={ 'Sign in to connect to WordPress.com' } />
 			);
